Add files via upload
"Loan Prediction Using Regression" is a data science project that focuses on predicting whether a loan applicant is likely to be approved or denied based on various features and historical data. Regression analysis is employed in this project as it involves predicting a continuous outcome, such as the likelihood of loan approval, based on input variables.